// apps/film/src/dto/creatFilm.dto.ts
export class CreatFilmDto {

  name: string;

  nameEng: string;

  rating: number;

  estimation: string;

  subtitles: string;

  audio_track: string;

  year: number;

  slogan: string;

  budget: string;

  tax_usa: string;

  tax_world: string;

  tax_rus: string;

  release_world: string;

  release_rus: string;

  release_dvd: string;

  age_restrict: string;

  time_during: string;

  rating_MPAA: string;

  trailer: string;

  picture_URL: string;

  picture_film: string;

  country: string[];
  genre: string[];
  filmDirector: string[];
  filmWriter: string[];
  filmProducer: string[];
  filmOperator: string[];
  filmComposer: string[];
  filmArtist: string[];
  filmEditor: string[];
  mainActors: string[];
  filmTranslators: string[];
  similarFilms: string[];


}